Update interpreter
New nodes: Program, Block, Type. VarsDeclaration;
Using BigDecimals everywhere;
Now base operations with real numbers are possible;

+++ src/ru/zelark/spi/interpreter/SymbolTable.java
@@ -1,7 +1,9 @@
 package ru.zelark.spi.interpreter;
 
+import java.math.BigDecimal;
+
 public interface SymbolTable {
-    Integer get(String name);
+    BigDecimal get(String name);
 
     void put(String name, Object value);
